+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New flag to add prefix to sales orders, invoices, credit memos and shipments.
+- Ability to change installed version when installing.
+
+### Changed
+- Call to the site url changed and backup added.
+
 
 ## [1.6.0] - 2020-10-16
 ### Added